Add the GitHub issue template (#668)
Collecting all the relevant details early on means less back and forth
between the poster and project contributors. Asking to create a small
reproducible demo means less time wasted trying to repro the bug.

Brunch has been doing this for a while and it has really helped.
There are people who deleted the template and went free-form in their
bug reports but for the most part, people have actually been filling
that in:
https://github.com/brunch/brunch/issues?q=is%3Aissue+is%3Aclosed

A lesser percentage of the people were creating the demos, but a few
definitely were. The details, and the demos especially, led to faster
resolution times, which a win for everyone.